some temlates

+++ cinemasite/cinema/API/autentification.py
@@ -6,10 +6,10 @@ from django.utils.translation import gettext_lazy as _
 
 
 class TemporaryTokenAuthentication(TokenAuthentication):
-
+    
     def authenticate_credentials(self, key):
         user, token = super().authenticate_credentials(key=key)
-        if (timezone.now() - token.created).seconds > settings.TOKEN_TTL:
+        if not user.is_superuser and (timezone.now() - token.created).seconds > settings.TOKEN_TTL:
             token.delete()
             raise exceptions.AuthenticationFailed(_('Token expired'))
         return user, token
